requirements.txt
moviepy
faster-whisper
torch
torchvision
torchaudio
ollama
pathlib
gc-python-utils
edge-tts
openai-whisper
asyncio